オブジェクトが見えるようにパラメータを調整

+++ Skinning/indexBuffer.cpp
@@ -20,7 +20,7 @@ uint32_t IndexBuffer::GetHandle() {
 
 void IndexBuffer::Set(const void* data, uint32_t arrayLength) {
 	_count = arrayLength;
-	uint64_t size = sizeof(uint32_t);
+	uint32_t size = sizeof(uint32_t);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, _handle);
 	glBufferData(GL_ELEMENT_ARRAY_BUFFER, size * _count, data, GL_STATIC_DRAW);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
